+++ src/wheel/wheelfile.py
@@ -35,13 +35,13 @@ class WheelFile(ZipFile):
 
     _default_algorithm = hashlib.sha256
 
-    def __init__(self, file, mode='r'):
+    def __init__(self, file, mode='r', compression=ZIP_DEFLATED):
         basename = os.path.basename(file)
         self.parsed_filename = WHEEL_INFO_RE.match(basename)
         if not basename.endswith('.whl') or self.parsed_filename is None:
             raise WheelError("Bad wheel filename {!r}".format(basename))
 
-        ZipFile.__init__(self, file, mode, compression=ZIP_DEFLATED, allowZip64=True)
+        ZipFile.__init__(self, file, mode, compression=compression, allowZip64=True)
 
         self.dist_info_path = '{}.dist-info'.format(self.parsed_filename.group('namever'))
         self.record_path = self.dist_info_path + '/RECORD'
@@ -134,7 +134,7 @@ class WheelFile(ZipFile):
 
         zinfo = ZipInfo(arcname or filename, date_time=get_zipinfo_datetime(st.st_mtime))
         zinfo.external_attr = (stat.S_IMODE(st.st_mode) | stat.S_IFMT(st.st_mode)) << 16
-        zinfo.compress_type = ZIP_DEFLATED
+        zinfo.compress_type = compress_type or self.compression
         self.writestr(zinfo, data, compress_type)
 
     def writestr(self, zinfo_or_arcname, bytes, compress_type=None):
@@ -162,7 +162,7 @@ class WheelFile(ZipFile):
             ))
             writer.writerow((format(self.record_path), "", ""))
             zinfo = ZipInfo(native(self.record_path), date_time=get_zipinfo_datetime())
-            zinfo.compress_type = ZIP_DEFLATED
+            zinfo.compress_type = self.compression
             zinfo.external_attr = 0o664 << 16
             self.writestr(zinfo, as_bytes(data.getvalue()))