+++ docs/news.rst
@@ -6,6 +6,7 @@ Release Notes
 - Dropped Python 3.4 support
 - Added automatic platform tag detection for macOS binary wheels
   (PR by Grzegorz Bokota)
+- Added the ``--compression=`` option to the ``bdist_wheel`` command
 - Fixed PyPy tag generation to work with the updated semantics (#328)
 - Updated project packaging and testing configuration for :pep:`517`
 - Moved the contents of setup.py to setup.cfg